Gen 3 - README files updates, clippy lints

+++ dis-assemble-gen-3/src/constants.rs
@@ -1,2 +1,2 @@
-pub const PDU_HEADER_LEN_BYTES: u16 = 16;
+pub const PDU_HEADER_LEN_BYTES: u8 = 16;
 pub const NO_REMAINDER: usize = 0;
